+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Add yearly view (#134)
+- Add internationalization(Spanish, French, German, and Chinese) i18n (#135)
 
 ## 2.2.0 (2023-10-06)